+++ src/main/java/com/lr/salestax/goods/impl/MiscGoods.java
@@ -21,7 +21,7 @@ public class MiscGoods
 	 * @param noOfItems
 	 * @param isImported
 	 */
-	public MiscGoods(final String goodsDescription, final double goodsCostPerItem, final double noOfItems, final boolean isImported) {
+	public MiscGoods(final String goodsDescription, final double goodsCostPerItem, final int noOfItems, final boolean isImported) {
 		super(goodsDescription, goodsCostPerItem, noOfItems, isImported);
 	}
